+++ contracts/facets/UniSwapFacet.sol
@@ -31,7 +31,7 @@ ERC2771ContextUpgradeable(address(forwarder))
  error COULD_NOT_PROCESS();
     /* Uniswap Router Address with interface*/
      function executeSwap(
-        uint24 _id,
+        uint256 _id,
         uint256 _oracleprice,
         uint24 poolfee
     ) external {
